finish css part of home bottom
