--- client/src/components/home.css
@@ -1,60 +1,0 @@
-body {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  line-height: 1.7;
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  width: 100svw;
-  min-height: 100svh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.alex-part {
-  display: grid;
-  place-content: center;
-}
-
-.weather-cross {
-  border: 1px solid #7bb5db;
-  background: rgb(248, 246, 246);
-  background: linear-gradient(0deg, rgb(253, 252, 252) 20%, rgb(207, 213, 217), rgb(172, 195, 209), rgb(141, 184, 211));
-  width: 100svw;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  gap: 1rem;
-  padding: 1rem;
-  font-size: 0.8rem;
-}
-
-.weather > p {
-  color: #0e0c5e;
-  font-weight: 800;
-}
-
-.weather > p > span {
-  color: #7bb5d9;
-  font-weight: 800;
-  margin-right: 2rem;
-}
-
-.weather-left {
-  display: flex;
-  flex-direction: column;
-}
-
-.weather-bottom {
-  display: flex;
-  flex-direction: column;
-}
-
-
-.separator-bottom {
-  border: 1px solid #7bb5db;
-}
-
-.wrap-chart {
-  max-width: 80svw;
-}
