+++ client/src/style/home.css
@@ -1,0 +1,60 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+
+body {
+
+  line-height: 1.4;
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  min-height: 100svh;
+
+}
+
+
+
+
+
+.weather-cross {
+  background: rgb(248, 246, 246);
+  background: linear-gradient(0deg, rgb(253, 252, 252) 20%, rgb(207, 213, 217), rgb(172, 195, 209), rgb(141, 184, 211));
+  width: 100vw;
+  padding: 1rem;
+  color: #0e0c5e;
+}
+
+
+
+/* COLUMN CSS */
+.weather-cross-col {
+  background: rgb(248, 246, 246);
+  background: linear-gradient(0deg, rgb(253, 252, 252) 20%, rgb(207, 213, 217), rgb(198, 211, 219), rgb(178, 208, 226));
+  width: 100vw;
+  font-size: 0.9rem;
+  display: flex;
+  justify-content: center;
+
+}
+
+.weather-data-col {
+  display: flex;
+  padding: 0.5rem 0.5rem;
+  gap: 4rem;
+}
+
+.data-col {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 1rem;
+}
+
+.data-col:first-child {
+  border-right: 2px solid #7bb5d9;
+  margin-bottom: 3rem;
+}
+
+.weather-data-col:first-child {
+  border-bottom: 2px solid #7bb5d9;
+}
